Create package `testenv` for test entities.
