+++ testutil-server/src/test/java/io/spine/testing/server/entity/EntityBuilderTest.java
@@ -27,9 +27,9 @@ import io.spine.base.Time;
 import io.spine.server.entity.Entity;
 import io.spine.server.entity.InvalidEntityStateException;
 import io.spine.testing.server.User;
-import io.spine.testing.server.entity.EntityBuilderTestEnv.TestEntity;
-import io.spine.testing.server.entity.EntityBuilderTestEnv.TestEntityBuilder;
-import io.spine.testing.server.entity.EntityBuilderTestEnv.UserAggregate;
+import io.spine.testing.server.entity.testenv.ebuilder.TestEntity;
+import io.spine.testing.server.entity.testenv.ebuilder.TestEntityBuilder;
+import io.spine.testing.server.entity.testenv.ebuilder.UserAggregate;
 import io.spine.validate.ConstraintViolation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
