+++ testutil-server/src/test/java/io/spine/testing/server/entity/EntitySubjectTest.java
@@ -25,8 +25,8 @@ import io.spine.server.entity.Entity;
 import io.spine.testing.server.SubjectTest;
 import io.spine.testing.server.blackbox.BbProjectId;
 import io.spine.testing.server.blackbox.BbProjectView;
-import io.spine.testing.server.entity.EntitySubjectTestEnv.ProjectView;
 import io.spine.testing.server.entity.given.Given;
+import io.spine.testing.server.entity.testenv.esubject.ProjectView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
